# server/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Voxmation API",
    "description": "Comprehensive API documentation for Voxmation - Voice automation and campaign management platform",
    "version": "1.0.0",
    "contact": {
      "name": "Voxmation Support",
      "email": "support@voxmation.com",
      "url": "https://voxmation.com"
    },
    "license": {
      "name": "proprietary",
      "url": "https://voxmation.com/license"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3001",
      "description": "Development server"
    },
    {
      "url": "https://api.voxmation.com",
      "description": "Production server"
    }
  ],
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT token obtained from /api/auth/login"
      },
      "ApiKey": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key",
        "description": "API key for programmatic access"
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "description": "Error code or message"
          },
          "message": {
            "type": "string",
            "description": "Detailed error description"
          }
        },
        "required": ["error"]
      },
...
      "Automation": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Automation UUID"
          },
          "account_id": {
            "type": "string"
          },
          "name": {
            "type": "string",
            "description": "Automation name"
          },
          "description": {
            "type": "string"
          },
          "type": {
            "type": "string",
            "enum": ["drip", "trigger", "welcome", "abandoned_cart", "re_engagement", "custom"],
            "description": "Automation type"
          },
          "trigger_type": {
            "type": "string",
            "description": "Trigger type (contact_tag, contact_property, etc.)"
          },
          "trigger_conditions": {
            "type": "object",
            "description": "Conditions that trigger the automation"
          },
          "workflow": {
            "type": "array",
            "description": "Workflow steps",
            "items": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "string"
                },
                "type": {
                  "type": "string",
                  "enum": ["send_email", "send_sms", "create_opportunity", "update_tag", "delay", "condition"]
                },
                "action": {
                  "type": "object"
                },
                "order": {
                  "type": "number"
                }
              }
            }
          },
          "status": {
            "type": "string",
            "enum": ["draft", "active", "paused", "completed", "archived"]
          },
          "is_recurring": {
            "type": "boolean"
          },
          "max_contacts_per_day": {
            "type": "number"
          },
          "total_contacts": {
            "type": "number"
          },
          "total_completed": {
            "type": "number"
          },
          "total_failed": {
            "type": "number"
          },
          "last_triggered_at": {
            "type": "string",
            "format": "date-time"
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": ["id", "name", "type", "trigger_type", "trigger_conditions", "workflow"]
      },
      "Call": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Call UUID"
          },
          "accountId": {
            "type": "string"
          },
          "twilio_call_sid": {
            "type": "string",
            "description": "Twilio call SID"
          },
          "to": {
            "type": "string",
            "description": "Recipient phone number"
          },
          "from": {
            "type": "string",
            "description": "Caller phone number"
          },
          "status": {
            "type": "string",
            "enum": ["queued", "ringing", "in-progress", "completed", "failed", "busy", "no-answer"],
            "description": "Call status"
          },
          "duration": {
            "type": "number",
            "description": "Call duration in seconds"
          },
          "startTime": {
            "type": "string",
            "format": "date-time"
          },
          "endTime": {
            "type": "string",
            "format": "date-time"
          },
          "recordingSid": {
            "type": "string",
            "description": "Recording SID if call was recorded"
          },
          "transcriptId": {
            "type": "string"
          },
          "transcript": {
            "type": "string",
            "description": "Call transcript (if available)"
          },
          "campaignId": {
            "type": "string",
            "description": "Associated campaign ID"
          },
          "metadata": {
            "type": "object",
            "additionalProperties": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": ["id", "to", "from", "status"]
      },
...
      "PaginatedResponse": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array"
          },
          "total": {
            "type": "number"
          },
          "page": {
            "type": "number"
          },
          "limit": {
            "type": "number"
          }
        }
      }
    }
  },
  "security": [
    {
      "BearerAuth": []
    }
  ],
  "paths": {
...
    "/api/automations": {
      "get": {
        "tags": ["Automations"],
        "summary": "List automations",
        "description": "Get all automations for the account",
        "operationId": "listAutomations",
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "description": "Filter by status",
            "schema": {
              "type": "string",
              "enum": ["draft", "active", "paused", "completed", "archived"]
            }
          },
          {
            "name": "type",
            "in": "query",
            "description": "Filter by type",
            "schema": {
              "type": "string",
              "enum": ["drip", "trigger", "welcome", "abandoned_cart", "re_engagement", "custom"]
            }
          },
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of automations",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedResponse"
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Automations"],
        "summary": "Create automation",
        "description": "Create a new automation (requires admin or marketing role)",
        "operationId": "createAutomation",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "example": "Welcome Series"
                  },
                  "description": {
                    "type": "string"
                  },
                  "type": {
                    "type": "string",
                    "enum": ["drip", "trigger", "welcome", "abandoned_cart", "re_engagement", "custom"]
                  },
                  "trigger_type": {
                    "type": "string",
                    "example": "contact_created"
                  },
                  "trigger_conditions": {
                    "type": "object",
                    "description": "Conditions that trigger the automation"
                  },
                  "workflow": {
                    "type": "array",
                    "description": "Array of workflow steps",
                    "items": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string"
                        },
                        "type": {
                          "type": "string",
                          "enum": ["send_email", "send_sms", "create_opportunity", "update_tag", "delay", "condition"]
                        },
                        "action": {
                          "type": "object"
                        },
                        "order": {
                          "type": "number"
                        }
                      },
                      "required": ["type", "action", "order"]
                    },
                    "minItems": 1
                  },
                  "is_recurring": {
                    "type": "boolean"
                  },
                  "max_contacts_per_day": {
                    "type": "number"
                  },
                  "tags": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  }
                },
                "required": ["name", "type", "trigger_type", "trigger_conditions", "workflow"]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Automation created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Automation"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/calls": {
      "get": {
        "tags": ["Calls"],
        "summary": "List calls",
        "description": "Get all calls for the account",
        "operationId": "listCalls",
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "description": "Filter by call status",
            "schema": {
              "type": "string",
              "enum": ["queued", "ringing", "in-progress", "completed", "failed", "busy", "no-answer"]
            }
          },
          {
            "name": "campaignId",
            "in": "query",
            "description": "Filter by campaign",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of calls",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedResponse"
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Calls"],
        "summary": "Initiate call",
        "description": "Make a new outbound call",
        "operationId": "createCall",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "to": {
                    "type": "string",
                    "description": "Recipient phone number",
                    "example": "+14155552671"
                  },
                  "from": {
                    "type": "string",
                    "description": "Caller phone number"
                  },
                  "campaignId": {
                    "type": "string"
                  },
                  "twimlUrl": {
                    "type": "string",
                    "format": "uri",
                    "description": "TwiML URL for call logic"
                  },
                  "record": {
                    "type": "boolean",
                    "description": "Record the call"
                  },
                  "recordingChannels": {
                    "type": "string",
                    "enum": ["mono", "stereo", "both"]
                  },
                  "statusCallback": {
                    "type": "string",
                    "format": "uri",
                    "description": "Webhook URL for status updates"
                  },
                  "metadata": {
                    "type": "object",
                    "additionalProperties": {
                      "type": "string"
                    }
                  }
                },
                "required": ["to"]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Call initiated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Call"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}